feat(providers): add tencent qq provider
